pyproject.toml
[tool.poetry]
name = "task-management-service"
version = "0.1.0"
description = "A FastAPI task management service with PostgreSQL"
authors = ["Azeasy <azeasy19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = "^0.27"
pydantic = {extras = ["email"], version = "^2.10.6"}
sqlalchemy = "^2.0"
asyncpg = "^0.29"
alembic = "^1.13"
python-jose = "^3.3"
bcrypt = "^4.0"
pydantic-settings = "^2.8.1"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"
fakeredis = "^2.27.0"
trio = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
